fixes extraction of client_id/secret in alpine
A recent Grafana upgrade appears to have switched the image to an alpine
base.

Alpine base only has the "diet" BusyBox version of tools like "grep"

the diet version of grep does not support the perl regex matching groups
and so was unable to extract the correct value from for the oauth client
id/secret from the big blob of JSON that is VCAP_SERVICES, causing the
authentication to be incorrectly setup.

This change switches to performing the task using sed, which while still
basically just as hacky as before, is able to perform the value
extraction required.